+++ packages/expo-camera/android/src/main/java/expo/modules/camera/ExpoCameraView.kt
@@ -719,6 +719,7 @@ class ExpoCameraView(
           target = id,
           data = barcode.value.toString(),
           raw = barcode.raw.toString(),
+          rawBytes = barcode.rawBytes,
           type = BarcodeType.mapFormatToString(barcode.type),
           cornerPoints = cornerPoints,
           bounds = boundingBox,